offline-jobs/requirements.txt
apache-airflow
confluent-kafka
cryptography
implicit
pandas
requests